Adding FilledInterpolation
A FilledInterpolation type is proposed which provides capabilities for a user-defined value to be used for filling the extrapolation domain of an Interpolation.  The FilledInterpolation type includes an Interpolation object that is constructed with any existing ExtrapolationBehavior.  The ExtrapolationBehavior of the Interpolation object is replaced with the user-defined fillvalue in the extrapolation region of the FilledInterpolation.

The approach taken for this functionality was proposed to me by @tlycken in response to my email inquiries about adding functionality similar to the use of an "extrapval" argument in Matlab's interp1/interp2/interpn functions and the fill_value keyword in scipy.interpolate's interp1d/interp2d/interpn functions.
Conflicts:
	src/Interpolations.jl