Improve the speed of filtering
This isn't a huge improvement, but in my tests it allocates about
seven-fold less memory and runs in approximately 60% of the
time of the original. The improvement is much more dramatic for
cases where it doesn't have to use Woodbury, in which case it can be
more than 5x faster.

This also turns off pivoting to improve the cache behavior.