Make itp inferrable, and test it
The return type of interpolate() was not inferrable, which would have
led to significant performance penalties when used without a
function barrier.

This also makes `size` inferrable, which is helpful because its
output is often used for array creation, and the created array's
type (dimensionality parameter) would not be inferrable.