Add a `pad` parameter to BSplineInterplation and fix `size`
Formerly, the output of size included padding, which I suspect
is not how we want it to work. By including the padding in the type,
we can also generate more reliable code (see `define_indices` in
`b-splines/quadratic.jl`)

This also prepares the way for in-place interplation (which would use
no padding).