Refactor extrapolation, and add behaviors Periodic and Reflect
This is a complete refactor of the extrapolation functionality, that allows
for much more flexible specifications of extrapolation behavior. It is now
possible to extrapolate with different schemes in different dimensions, as
well as in different directions in the same dimension (high/low values), and
any possible combinations thereof.

The framework is intended to be composable, so implementing new schemes should
only require code equivalent to that found in src/extrapolation/flat.jl