Corrections in the patch
Changes made to generalize the call and remove extraneous external castings. 
Additionally, an example was included for mixed dimensions, which does require the explicit type declaration.
Finally, a description of the NoInterp mode was added.

Note: An update will be needed to change the syntax according to the new changes in the NoInterp syntax.