Decouple extrapolation bounds from element type
Extrapolation bounds for gridded interpolations are effectively indices and not coupled with the array eltype.  So this changes it to use integers (for `OnGrid()`) and SimpleRatios (for `OnCell()`) directly, instead of trying to convert the values to the eltype.  This is important because some array values may not support conversion from integers or floats.