Remove the call to `real` when rounding to integer indexes
The deprecation of `real` for DualNumbers broke the tests that used `@inferrable`. More long-term, I think it's better to just perform the operation we want without an extra call to `real`. This places the onus on the relevant number types to support `round(Int, x)`, but the two major use-cases for automatic differentiation (ForwardDiff and DualNumbers) now support this.