Narrow compatibility requirements for StaticArrays
Currently, `using Interpolations` fails with StaticArrays 0.10 and 0.11
due to the missing `StaticArrays.SA` value. This PR removes the claimed
compatibility to avoid users accidentally getting a broken installation.

We'll also need to edit the General registry to apply this compatibility
fix to Interpolations.jl v0.13.0 and v0.13.1 otherwise the package
resolver will incorrectly want to downgrade to those versions if a user
has an older StaticArrays version.